# Suppress heap allocations in TracerSettings.IsIntegrationEnabled