- Remove closure allocation in IntegrationSettingsCollection
- Getting the AppDomain friendly name seems to internally call Assembly.GetName, which seems to allocate quite a bit (especially on .net core). Used AppDomain.SetData to cache the result and prevent further allocations

Benchmark:
```csharp
TracerSettings.IsIntegrationEnabled("Test")
```

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-VTGAFA : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-MNUBEA : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method |       Runtime |      Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-------------- |----------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|    Old |    .NET 4.7.2 | 254.26 ns |  5.100 ns |  7.149 ns |  1.00 |    0.00 | 0.0191 |     - |     - |     120 B |
|    New |    .NET 4.7.2 | 108.29 ns |  2.160 ns |  2.731 ns |  0.43 |    0.02 |      - |     - |     - |         - |
|        |               |           |           |           |       |         |        |       |       |           |
|    Old | .NET Core 3.1 | 894.16 ns | 17.648 ns | 28.499 ns |  1.00 |    0.00 | 0.0744 |     - |     - |     472 B |
|    New | .NET Core 3.1 |  59.63 ns |  1.178 ns |  1.102 ns |  0.07 |    0.00 |      - |     - |     - |         - |