@kevingosse 2020-06-30
Rewrote to use a static field + added a unit test (this bit of logic seems to critical not to have one).

New benchmark:

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-VTGAFA : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-MNUBEA : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method |       Runtime |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-------------- |----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
|    Old |    .NET 4.7.2 | 224.47 ns | 3.787 ns | 3.542 ns |  1.00 | 0.0191 |     - |     - |     120 B |
|    New |    .NET 4.7.2 | 103.12 ns | 0.888 ns | 0.831 ns |  0.46 |      - |     - |     - |         - |
| Static |    .NET 4.7.2 |  28.96 ns | 0.236 ns | 0.184 ns |  0.13 |      - |     - |     - |         - |
|        |               |           |          |          |       |        |       |       |           |
|    Old | .NET Core 3.1 | 820.41 ns | 7.249 ns | 6.053 ns |  1.00 | 0.0744 |     - |     - |     472 B |
|    New | .NET Core 3.1 |  56.87 ns | 0.438 ns | 0.388 ns |  0.07 |      - |     - |     - |         - |
| Static | .NET Core 3.1 |  25.99 ns | 0.159 ns | 0.149 ns |  0.03 |      - |     - |     - |         - |


I also realized, while writing the test, that the whole thing could be eluded on .net core (since there are no appdomains), but I'd say it's cheap enough not to bother for now.